created resize wrapper, removed resizing from cells
