+++ src/ModelTable/ModelTableHead.tsx
@@ -3,7 +3,8 @@ import { DndSortableWrapper, humanizeText } from '@/utils';
 
 import { ModelTableHeadMenu } from './ModelTableHeadMenu';
 import { useModelTableStore } from './useModelTableStore';
-import { useEffect, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
+import { ResizeWrapper } from '@/utils/components/ResizeWrapper';
 
 export interface ModelTableHeadProps extends Omit<TableHeadProps, 'columnId'> {
   field: string;
@@ -21,68 +22,28 @@ export const ModelTableHead = ({
   const draggable = useModelTableStore(
     (state) => state.tableOptions.draggable ?? true,
   );
+  const resizable = useModelTableStore(
+    (state) => state.tableOptions?.columnOptions?.[field]?.resizable ?? true
+  )
+  const fieldWidth = useModelTableStore(
+    (state) => state.tableOptions?.columnOptions?.[field]?.width ?? 0
+  )
 
   const cellRef = useRef<HTMLTableCellElement>(null);
-  const [width, setWidth] = useState<number|undefined>();
-  const [startX, setStartX] = useState<number>();
-  const [startWidth, setStartWidth] = useState<number>();
-  const [finishedResize, setFinishedResize] = useState<boolean>(false);
-
-  if (!width && cellRef.current) {
-    setWidth(cellRef.current.offsetWidth)
-  }
-
-  useEffect( () => {
-    if (startX && startWidth) {
-      document.addEventListener("mousemove", doResize);
-      document.addEventListener("mouseup", stopResize);
-    }
-  }, [startX, startWidth])
-
-  useEffect( () => {
-    if (finishedResize && cellRef?.current && width) {
-      setFinishedResize(false);
-      setWidth(Math.max(cellRef.current?.offsetWidth, width))
-    }
-  }, [finishedResize, width])
-
-  const startResizing = (
-    e : React.MouseEvent<HTMLDivElement, MouseEvent>
-  ) => {
-    setStartX(e.clientX);
-    setStartWidth(width);
-  }
-
-  const doResize = (e : MouseEvent) => {
-    if (startWidth && startX) {
-      console.log(startWidth + (e.clientX - startX))
-      setWidth(startWidth + (e.clientX - startX))
-    }
-  }
-
-  const stopResize = () => {
-    document.removeEventListener("mousemove",doResize);
-    document.removeEventListener("mouseup",stopResize);
-    setFinishedResize(true);
-  }
+  const [width, setWidth] = useState<number>(fieldWidth);
 
   return (
     <DndSortableWrapper draggable={draggable} dndId={field}>
       <TableHead columnId={field} {...tableHeadProps} ref={cellRef} style={{ width: width }}>
-        {children === undefined ? (
-          <div className='flex h-full flex-row'>
+        <ResizeWrapper cellRef={cellRef} width={width} setWidth={setWidth} resizable={resizable}>
+          {children === undefined ? (
             <ModelTableHeadMenu field={field}>
               {label ?? humanizeText(field)}
             </ModelTableHeadMenu>
-            <div className="flex-grow"/>
-            <div className='z-50 h-full w-2 cursor-ew-resize select-none bg-red-100' onMouseDown={(e) => {
-              startResizing(e)
-              e.stopPropagation()
-            }}/>
-          </div>
-        ) : (
-          children
-        )}
+          ) : (
+            children
+          )}
+        </ResizeWrapper>
       </TableHead>
     </DndSortableWrapper>
   );
